[#20241026] Split settings tabs into separate components
